requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
gunicorn==20.1.0
CairoSVG==2.7.1
Pillow==10.2.0
python-pptx==0.6.23